+++ src/components/orchestrator/Orchestrator.test.tsx
@@ -1,17 +1,29 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { useSearch } from '@tanstack/react-router'
 import { act, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { expect } from 'vitest'
 
 import { MODE_TYPE } from '@/constants/mode'
+import { PAGE_TYPE } from '@/constants/page'
 import { TELEMETRY_EVENT_TYPE } from '@/constants/telemetry'
 import { TelemetryContext } from '@/contexts/TelemetryContext'
+import type { Interrogation } from '@/models/interrogation'
 import { renderWithRouter } from '@/utils/tests'
 
 import { Orchestrator } from './Orchestrator'
 
+vi.mock('@tanstack/react-router', async (importOriginal) => {
+  const actual =
+    (await importOriginal()) as typeof import('@tanstack/react-router')
+  return {
+    ...actual,
+    useSearch: vi.fn(),
+  }
+})
+
 describe('Orchestrator', () => {
-  const interrogation = {
+  const defaultInterrogation = {
     stateData: undefined,
     data: undefined,
     personalization: undefined,
@@ -50,16 +62,21 @@ describe('Orchestrator', () => {
     maxPage: '2',
   }
   const queryClient = new QueryClient()
+
+  const mockGetDepositProof = vi.fn()
+
   const OrchestratorTestWrapper = ({
     mode,
+    initialInterrogation = defaultInterrogation,
   }: {
     mode: MODE_TYPE.COLLECT | MODE_TYPE.REVIEW | MODE_TYPE.VISUALIZE
+    initialInterrogation?: Interrogation
   }) => (
     <QueryClientProvider client={queryClient}>
       <Orchestrator
         metadata={metadata}
         mode={mode}
-        initialInterrogation={interrogation}
+        initialInterrogation={initialInterrogation}
         // @ts-expect-error: we should have a better lunatic mock
         source={source}
         getReferentiel={() => {
@@ -68,9 +85,7 @@ describe('Orchestrator', () => {
         updateDataAndStateData={() => {
           return new Promise<void>(() => {})
         }}
-        getDepositProof={() => {
-          return new Promise<void>(() => {})
-        }}
+        getDepositProof={mockGetDepositProof}
       />
     </QueryClientProvider>
   )
@@ -258,5 +273,34 @@ describe('Orchestrator', () => {
 
     act(() => getByText('Continue').click())
     expect(document.title).not.toContain('*')
+  })
+
+  it('calls getDepositProof with right params', async () => {
+    // Set initial interrogation as validated on end page to directly enable download receipt
+    const initialInterrogation: Interrogation = {
+      ...defaultInterrogation,
+      stateData: { state: 'VALIDATED', currentPage: PAGE_TYPE.END, date: 1 },
+    }
+
+    // Provide encoded param via URL
+    vi.mocked(useSearch).mockReturnValue({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
+
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper
+        mode={MODE_TYPE.COLLECT}
+        initialInterrogation={initialInterrogation}
+      />,
+    )
+
+    vi.mocked(mockGetDepositProof).mockResolvedValue(undefined)
+
+    // download deposit proof
+    act(() => getByText('Download the receipt').click())
+
+    expect(mockGetDepositProof).toHaveBeenCalledWith({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
   })
 })